Backend/requirements.txt
# --- FastAPI Core ---
fastapi==0.115.0
uvicorn[standard]==0.30.0

# --- HTTP Client ---
httpx==0.24.1

# --- Database (Async) ---
SQLAlchemy[asyncio]==2.0.34
psycopg[binary]==3.2.1

# --- Configuration ---
pydantic-settings==2.3.0
pydantic==2.9.2
email-validator==2.1.0

# --- Authentication ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# --- Testing ---
pytest>=8.0.0,<9.0.0
pytest-asyncio==0.23.8

# --- Video transcripts ---
youtube-transcript-api>=1.2.4,<2.0

# --- Env / HTTP utilities ---
python-dotenv>=1.0.0
requests>=2.31.0
certifi

# --- Postgres vector support ---
pgvector==0.2.5